OSIS-908 Use the same view for attribution and update_learning_unit_pedagogy
