+++ base/templates/learning_unit/pedagogy.html
@@ -1,7 +1,6 @@
 {% extends "layout.html" %}
 {% load staticfiles %}
 {% load i18n bootstrap3 %}
-{% load getattr %}
 
 {% comment "License" %}
 * OSIS stands for Open Student Information System. It's an application
@@ -37,39 +36,11 @@
     <div class="panel-body" id="tabs" >
         {% include "learning_unit/blocks/tabs.html" %}
         <div class="tab-content" id="tab_content">
-            <div role="tabpanel" class="tab-pane active" id="pedagogy">
-                <form method="post">
-                    {% csrf_token %}
-                    {% bootstrap_field summary_editable_form.summary_locked field_class="well"%}
-
-                    {% include "learning_unit/blocks/panel/educational_informations.html" with can_edit_information=perms.base.can_edit_learningunit_pedagogy %}
-
-                    {% include "learning_unit/blocks/panel/bibliography.html" with formset=bibliography_formset %}
-                    {% bootstrap_field summary_editable_form.mobility_modality %}
-                    {% buttons %}
-                        {% bootstrap_button _("save") button_type="submit" button_class="btn-primary pull-right" %}
-                    {% endbuttons %}
-                </form>
-            </div>
+            {% url 'learning_unit_pedagogy_edit' learning_unit_year_id=learning_unit_year.id as post_url %}
+            {% include "learning_unit/blocks/panel/educational_informations.html" with can_edit_information=perms.base.can_edit_learningunit_pedagogy %}
         </div>
     </div>
 </div>
 
 <div class="modal fade" id="pedagogy_edit" role="dialog" data-backdrop="static"></div>
-{% endblock %}
-
-{% block script %}
-    <script src="{% static 'js/pedagogy_modal.js' %}"></script>
-
-    <script src="{% url 'javascript-catalog' %}"></script>
-    <script src="{% static 'js/jquery.formset.js' %}"></script>
-    <script>
-    $('.link-formset').formset({
-        addText: gettext('add another'),
-        deleteText: gettext('remove'),
-        addCssClass: 'btn btn-info btn-sm',
-        deleteCssClass: 'btn btn-warning btn-sm',
-        prefix: 'bibliography_set'
-    });
-    </script>
 {% endblock %}
