+++ base/tests/views/test_learning_unit.py
@@ -84,7 +84,8 @@ from base.tests.factories.person import PersonFactory
 from base.tests.factories.person_entity import PersonEntityFactory
 from base.tests.factories.user import SuperUserFactory, UserFactory
 from base.views.learning_unit import compute_partim_form_initial_data, _get_post_data_without_read_only_field, \
-    learning_unit_components, learning_class_year_edit, learning_unit_pedagogy, learning_unit_specifications
+    learning_unit_components, learning_class_year_edit, learning_unit_specifications
+from base.views.learning_units.update import learning_unit_pedagogy
 from base.views.learning_units.search import learning_units_service_course
 from cms.enums import entity_name
 from cms.tests.factories.text_label import TextLabelFactory
